+++ grib2/pngpack.c
@@ -73,7 +73,7 @@ void pngpack(g2float *fld,g2int width,g2int height,g2int *idrstmpl,
         if (fld[j] > rmax) rmax=fld[j];
         if (fld[j] < rmin) rmin=fld[j];
       }
-      maxdif = (g2int)rint( (rmax-rmin)*dscale*bscale );
+      maxdif = (g2int)rint( ((double)rmax-rmin)*dscale*bscale );
 /*
 //  If max and min values are not equal, pack up field.
 //  If they are equal, we have a constant field, and the reference
@@ -91,15 +91,15 @@ void pngpack(g2float *fld,g2int width,g2int height,g2int *idrstmpl,
            //  No binary scaling and calculate minumum number of 
            //  bits in which the data will fit.
            */
-           imin=(g2int)rint(rmin*dscale);
-           imax=(g2int)rint(rmax*dscale);
+           imin=(g2int)rint((double)rmin*dscale);
+           imax=(g2int)rint((double)rmax*dscale);
            maxdif=imax-imin;
            temp=log((double)(maxdif+1))/alog2;
            nbits=(g2int)ceil(temp);
            rmin=(g2float)imin;
            /*   scale data*/
            for(j=0;j<ndpts;j++)
-             ifld[j]=(g2int)rint(fld[j]*dscale)-imin;
+             ifld[j]=(g2int)rint(fld[j]*(double)dscale)-imin;
         }
         else {
            /*
@@ -108,12 +108,12 @@ void pngpack(g2float *fld,g2int width,g2int height,g2int *idrstmpl,
            */
            rmin=rmin*dscale;
            rmax=rmax*dscale;
-           maxdif=(g2int)rint((rmax-rmin)*bscale);
+           maxdif=(g2int)rint(((double)rmax-rmin)*bscale);
            temp=log((double)(maxdif+1))/alog2;
            nbits=(g2int)ceil(temp);
            /*   scale data*/
            for (j=0;j<ndpts;j++)
-             ifld[j]=(g2int)rint(((fld[j]*dscale)-rmin)*bscale);
+             ifld[j]=(g2int)rint(((fld[j]*(double)dscale)-rmin)*bscale);
         }
         /*
         //  Pack data into full octets, then do PNG encode.
